♻️ refactor: extract store/ to top-level package
- Move agent/store/ → store/
- Update imports in agent/, cmd/anna/
- Store is a cross-cutting concern, not agent-specific
