+++ store/index_test.go
